# docs/swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "A part of microservice infrastructure, who responsible for customer user entity.",
    "title": "Customer Service",
    "contact": {
      "name": "Artem Kostenko",
      "url": "https://github.com/aerosystems"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.1"
  },
  "host": "gw.verifire.dev/customer",
  "basePath": "/",
  "paths": {
    "/v1/firebase/create-customer": {
      "post": {
        "description": "Create user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Create user",
        "parameters": [
          {
            "description": "Create user",
            "name": "raw",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HTTPServer.CreateFirebaseCustomerRequestBody"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HTTPServer.ErrorResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/HTTPServer.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/HTTPServer.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "HTTPServer.CreateFirebaseCustomerRequestBody": {
      "type": "object",
      "properties": {
        "message": {
          "type": "object",
          "properties": {
            "data": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          }
        },
        "subscription": {
          "type": "string"
        }
      }
    },
    "HTTPServer.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Should contain Access JWT Token, with the Bearer started",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}